Extract data to repository
