+++ src/main/java/com/example/kitchenproject/RecipeRepository.java
@@ -1,0 +1,11 @@
+package com.example.kitchenproject;
+
+import com.example.kitchenproject.model.Recipe;
+
+import java.util.List;
+
+public interface RecipeRepository {
+    List<Recipe> getAllRecipes();
+
+    List<String> getAllIngredientNames();
+}
